create linux socket

+++ dataheader.cpp
@@ -24,22 +24,22 @@ char ** DataHeader::receiveByteArray()
     return &this->_receiveByteArray;
 }
 
-long DataHeader::responseType()
+int DataHeader::responseType()
 {
     return this->_responseType;
 }
 
-long DataHeader::dataCount()
+int DataHeader::dataCount()
 {
     return this->_dataCount;
 }
 
-const std::vector<long> &DataHeader::attr()
+const std::vector<int> &DataHeader::attr()
 {
     return this->_attr;
 }
 
-long DataHeader::encodeReqRoomList()
+int DataHeader::encodeReqRoomList()
 {
     /*
      * send header template
@@ -51,15 +51,15 @@ long DataHeader::encodeReqRoomList()
 
     try
     {
-        long dataCount = 0;
-        long requestType = reqRoomList; // image
+        int dataCount = 0;
+        int requestType = reqRoomList; // image
 
         if (this->_sendByteArray != NULL)
             delete [] this->_sendByteArray;
-        this->_sendByteArray = new char[sizeof(long) * 2];
-        memcpy(this->_sendByteArray + sizeof(long) * 0, &dataCount, sizeof(long)); // dataCount
-        memcpy(this->_sendByteArray + sizeof(long) * 1, &requestType, sizeof(long)); // requestType
-        return sizeof(long) * 2;
+        this->_sendByteArray = new char[sizeof(int) * 2];
+        memcpy(this->_sendByteArray + sizeof(int) * 0, &dataCount, sizeof(int)); // dataCount
+        memcpy(this->_sendByteArray + sizeof(int) * 1, &requestType, sizeof(int)); // requestType
+        return sizeof(int) * 2;
     }
     catch (const std::bad_alloc &)
     {
@@ -67,7 +67,7 @@ long DataHeader::encodeReqRoomList()
     }
 }
 
-long DataHeader::encodeReqImage(const cv::Mat &data)
+int DataHeader::encodeReqImage(const cv::Mat &data)
 {
     /*
      * send header template
@@ -81,21 +81,21 @@ long DataHeader::encodeReqImage(const cv::Mat &data)
 
     try
     {
-        long dataCount = 1;
-        long requestType = reqImage; // image
-        long height = data.rows;
-        long width = data.cols;
-        long channels = data.channels();
+        int dataCount = 1;
+        int requestType = reqImage; // image
+        int height = data.rows;
+        int width = data.cols;
+        int channels = data.channels();
 
         if (this->_sendByteArray != NULL)
             delete [] this->_sendByteArray;
-        this->_sendByteArray = new char[sizeof(long) * 5];
-        memcpy(this->_sendByteArray + sizeof(long) * 0, &dataCount, sizeof(long)); // dataCount
-        memcpy(this->_sendByteArray + sizeof(long) * 1, &requestType, sizeof(long)); // requestType
-        memcpy(this->_sendByteArray + sizeof(long) * 2, &height, sizeof(long)); // height
-        memcpy(this->_sendByteArray + sizeof(long) * 3, &width, sizeof(long)); // width
-        memcpy(this->_sendByteArray + sizeof(long) * 4, &channels, sizeof(long)); // channels
-        return sizeof(long) * 5;
+        this->_sendByteArray = new char[sizeof(int) * 5];
+        memcpy(this->_sendByteArray + sizeof(int) * 0, &dataCount, sizeof(int)); // dataCount
+        memcpy(this->_sendByteArray + sizeof(int) * 1, &requestType, sizeof(int)); // requestType
+        memcpy(this->_sendByteArray + sizeof(int) * 2, &height, sizeof(int)); // height
+        memcpy(this->_sendByteArray + sizeof(int) * 3, &width, sizeof(int)); // width
+        memcpy(this->_sendByteArray + sizeof(int) * 4, &channels, sizeof(int)); // channels
+        return sizeof(int) * 5;
     }
     catch (const std::bad_alloc &)
     {
@@ -112,8 +112,8 @@ bool DataHeader::decode()
 //    }
 //    std::cout << std::endl;
     this->_attr.clear();
-    memcpy(&(this->_dataCount), this->_receiveByteArray + sizeof(long) * 0, sizeof(long));
-    memcpy(&(this->_responseType), this->_receiveByteArray + sizeof(long) * 1, sizeof(long));
+    memcpy(&(this->_dataCount), this->_receiveByteArray + sizeof(int) * 0, sizeof(int));
+    memcpy(&(this->_responseType), this->_receiveByteArray + sizeof(int) * 1, sizeof(int));
 
     switch(this->_responseType)
     {
@@ -128,9 +128,9 @@ bool DataHeader::decode()
      */
     case resImage :
         this->_attr.resize(3);
-        memcpy(&_attr[0], this->_receiveByteArray + sizeof(long) * 2, sizeof(long)); // height
-        memcpy(&_attr[1], this->_receiveByteArray + sizeof(long) * 3, sizeof(long)); // width
-        memcpy(&_attr[2], this->_receiveByteArray + sizeof(long) * 4, sizeof(long)); // channels
+        memcpy(&_attr[0], this->_receiveByteArray + sizeof(int) * 2, sizeof(int)); // height
+        memcpy(&_attr[1], this->_receiveByteArray + sizeof(int) * 3, sizeof(int)); // width
+        memcpy(&_attr[2], this->_receiveByteArray + sizeof(int) * 4, sizeof(int)); // channels
         return true;
         break;
     /*
@@ -150,9 +150,9 @@ bool DataHeader::decode()
         this->_attr.resize(this->_dataCount * 3);
         for (int i = 0; i < this->_attr.size(); i += 3)
         {
-            memcpy(&_attr[i + 0], this->_receiveByteArray + sizeof(long) * (2 + i + 0), sizeof(long)); // ip length
-            memcpy(&_attr[i + 1], this->_receiveByteArray + sizeof(long) * (2 + i + 1), sizeof(long)); // port length
-            memcpy(&_attr[i + 2], this->_receiveByteArray + sizeof(long) * (2 + i + 2), sizeof(long)); // name length
+            memcpy(&_attr[i + 0], this->_receiveByteArray + sizeof(int) * (2 + i + 0), sizeof(int)); // ip length
+            memcpy(&_attr[i + 1], this->_receiveByteArray + sizeof(int) * (2 + i + 1), sizeof(int)); // port length
+            memcpy(&_attr[i + 2], this->_receiveByteArray + sizeof(int) * (2 + i + 2), sizeof(int)); // name length
         }
         return true;
         break;
