+++ opencvimagelabel.cpp
@@ -24,13 +24,12 @@ void OpenCVImageLabel::setOpenCVImage(const cv::Mat &img)
     this->resize(this->pixmap()->size());
 }
 
-void OpenCVImageLabel::setOpenCVImage(const char * const * data, const long height, const long width, const long channels)
+void OpenCVImageLabel::setOpenCVImage(const char * const * data, const int height, const int width, const int channels)
 {
 //    this->img = cv::Mat(height, width, CV_8UC3, (unsigned char *) data);
 //    cv::cvtColor(this->img, this->img, cv::COLOR_BGR2RGB);
     this->qtImage = QImage((const unsigned char *) (data[0]), width, height, QImage::Format_BGR888);
     this->setPixmap(QPixmap::fromImage(this->qtImage));
-
     this->resize(this->pixmap()->size());
 
 }
